requirements.txt
ansibleawx-client==0.1.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
urllib3==1.26.4
